Fix qtlocation build for WEC7.
Apparently math.h in WEC7 does not define M_PI used by qtlocation.
Added the requiresd define the necessary files.

qSqrt returns qreal and it is defined to 'float' on some configurations
and to 'double' in other configurations. In those configirations where
it is defined to float, no suitable overload for division is found.
Added explicit cast for qSqrt return value to double to fix the build.

Change-Id: I734fab66abfd4ce95e02ec83e373f0eacd87bdef
Reviewed-by: Andreas Holzammer <andreas.holzammer@kdab.com>
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>