PositionSource property fixes.
PositionSource has a couple of property signal issues which stem from
the fact that the order of property evaluation is not well defined
during component construction.  Fix emission of signals when the
underlying position info source changes.  When a new position info
source is created the updatInterval, preferredPositioningMethods and
supportedPositioningMethods change notification signals are emitted if
necessary.

Construction of a default position info source is delayed until after
the component has been constructed.  This saves creating a default
position info source in the constructor and then (possibly) replacing
it when the name property is set.

Fix possibility of setting unsupported position methods as preferred.
When setting the preferred positioning methods call the base class
implementation first, to ensure that only supported positioning methods
are set, prior to acting on the new preferred positioning methods.

Change-Id: Ie402f87ca7d82b52fd4678c40867894d3b122942
Reviewed-by: Alex <alexander.blasche@digia.com>