Fix recommendations search in example
The recommendations search did not work because a search area
was being passed along with a recommenation id.

Modify the example to only pass a recommendation id by itself.

The PlaceSearchModel had to be updated as well so that whenever a
null searchArea was assigned, the search request would send
a default QGeoShape as a search area, as opposed to not being
updated at all.

Change-Id: Ia55b2d74e7ac770412fb35e295ee07178cfaeb46
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>