Fix incorrect warning messages + refactor error handling
There was an assumption in the QML models that if the
QGeoServiceProvider::placeManager() returned null, then places was
not supported.  This is no longer the case as incorrect parmeters
may result in null being returned.

Previously a warning was output to console.  Instead, if possible,
we set the error status of the model, and set the error string
to the one provided by the QGeoServiceProvider.  If it is not
possible, such as in the case of the Icon element, a warning is
output to console but only if it is deemed necessary.

Task-number: QTBUG-26566
Change-Id: I584222af9325ddb8575fa8f77909521c49bc850a
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>