Fix incorrect error code during details fetch for nokia plugin
Previously a CommunicationError was returned if the id did not exist,
this was because an incorrect id would result in a non-existent URL
being used.  Now we make the assumption if the URL content is not
found, then it must be because the id did not exist.  In such
a case we can return a PlaceDoesNotExistError.

Change-Id: Ibfc21b826e0514fd694cb69d0372ef99e46e768e
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>