Fix incorrect error code during recommendation search
Previously a CommunicationError was returned if the id did not exist,
this was because an incorrect id would result in a non-esistent URL
being used.  Now we make the assumption that if the URL content is not
found, then it must be because the id did not exist.  In such a case
we can return a PlaceDoesNotExistError.

Change-Id: I50ddacb2341639dd6e227e795cefc5e92eea3d6e
Reviewed-by: Alex <alex.blasche@nokia.com>