Fix behaviour of update() of CategoryModel
update() should immediately go into the Loading state if an operation
is not already underway. Previously, if there was an error during
update(), the model would go straight into the Error state,
this patch modifies it so the behviour is Loading->Error.

Also add unit test to check update()'s behaviour when
encountering error conditions.

The behaviour is also modified such that setting
a plugin does not immediately trigger an update().
Explicitly calling update() is necessary to cause
a requery.

Change-Id: I98a328a251552aa00490a533552ec4b5f4509180
Reviewed-by: Aaron McCarthy <aaron.mccarthy@nokia.com>