Fix the QML list interface.
The QML list interface was changed, so that there is now a read only
constructor and the default properties are removed. This is because the
original interface was very runtime centric, like other parts in QML.
The new interface was changed for the needs of the QML designer and
debugger.

Task-number: QTBUG-9390
Change-Id: I1ed36c275fc7f8767956e55ad9afddcb293a970c
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@digia.com>