positionUpdated signal not emitted
The SerialPort returns (most of the times) only fragments of data when
being asked for a line by readLine(). This is due to fact, that it can
not know if the stream of data will continue and hence it will detect a
end of data. The QIODevice readLine() is specified to return data in
this situation, resulting in incomplete data and therefore incomplete
NMEA sentences, which can not be parsed by parsePosInfoFromNmeaData().
The canReadLine() function, which is used as a barriere in this fix,
only returns true if the available data of the SerialPort contains an
'\n' character. So only full lines will be read and parsed.

Task-number: QTMOBILITY-2054

Change-Id: I72eb1828fa850356322d4951a391ff568d022b78
Reviewed-by: Aaron McCarthy <mccarthy.aaron@gmail.com>