Ensure compilation does not fail if function does not return a value
In some cases the 3rd party project source tree "poly2tri" does contain
an "assert(0)" instead of a return statement. This lets at least the
GNU compiler assume a return value is missing.

Some systems consider this an error. To prevent compile errors on those
this change adds an additional compiler flag to make functions without a valid
return value a warning instead of an error.

Change-Id: I13acda8bfd274e37195d49392f4f93990f57ded2
Reviewed-by: Alex <ablasche@gmail.com>