Fix autotests while shadow building with install path
These tests require test plugins which must be installed. However
the CI does not run "make install" when shadow building with install path.
Subsequently the plugins are not found.

We add the plugin path of the shadow build dir to the library path
enabling the various plugin loader cases to find the required plugin
binaries.

Change-Id: I4df21278f3e28db01b3062b27ccc317ccbab3142
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Alex <alexander.blasche@digia.com>