Support explore functionality.
Add support for Nokia's explore functionality in places. This adds the
ProposedSearchResult search result type. The proposed search result has
a QPlaceSearchRequest, which can be submitted to
QPlaceManager::search() to perform the proposed search.

This has the added advantage that any arbitrary search request can be
returned as a search result. To facilitate this further a search
context field has been added to the place search request. This field
can be used by backends to store additional search context. Other
relevant fields should also be filled in. This allows additional search
context to be kept.

The Nokia v2 places implementation has be updated to use the explore
feature and to return ProposedSearchResults as well as place results.

Task-number: QTBUG-24874
Change-Id: Ib20c5a8c8a60e6bf16f38dcd75961a752b8b1b32
Reviewed-by: Alex <alexander.blasche@digia.com>