Add support for Geoclue Satellite interface.
The Tizen project introduced an interface for getting the Satellite
info directly from a positioning provider. If Geoclue supports this
interface make use of it in the geoclue position plugin.

The plugin no longer checks the gconf setting
/apps/geoclue/master/org.freedesktop.Geoclue.GPSDevice for a valid GPS
device. This setting is used by Gypsy (and other providers) but may not
necessarily be used by all providers that provide GPS quality data.
Instead rely on the Geoclue master provider to determine if GPS is
available. Because of this change the gconf dependency has been dropped.

The plugin no longer uses the Geoclue master provider as a proxy for
position data. Instead it queries the master provider for the DBus
service and path of the actual provider and connects directly to that
provider. This has to be done anyway to get velocity and satellite
data.

The plugin no longer assumes that the real positioning provider is
Gypsy. This assumption was hard coded as it was always connecting
directly to Gypsy to get the velocity data irrespective of what
positioning provider was actually being used. Geoclue master does not
proxy the Velocity interface. Instead it queries the master provider
for the DBus service and path of the actual provider and connects
directly to that provider. If the velocity interface is available it
will now be from the same provider as the position.

Change-Id: I8a6116e4061e50cc28d1efd71df0f7a348a0f143
Reviewed-by: Aaron McCarthy <mccarthy.aaron@gmail.com>