Fix Geoclue plugin when used with slow Geoclue providers.
The Geoclue plugin expected the master provider to select a positioning
provider quickly and would fail if one was not selected. This is likely
to happen if the only provider available is GPS based and has to obtain
a fix from a cold start.

Fixed by changing the Geoclue plugin to report success as long as the
master client interface is successfully created.

Change-Id: Ib74e3800a2444eb80afd26c671371de884759119
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>