Emit updates from Geoclue when an update interval is not specified.
An update interval of 0 is valid and means provide position updates as
frequently as the provider can supply them. The state of the update
timer was being used as the check for whether updates should be
provided. This is invalid for the case of a 0 update interval as the
update timer is not started.

Change-Id: I2c9d3fef41abaa21feafd31b232a7053d179ffd6
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>