Update position when switching position sources.
The declarative position source only updates its position when an update
arrives. The position property is now updated when:

At construction time the position source may have a last known position
available.

When switching position sources, the position is updated to the last
known position of the new source or invalidated in the case that no
position source is selected.

Change-Id: Ib67407f252f2e60d8fd4e32179847f465c1aaf94
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>