Removed timer for press time in QDeclarativeGeoMapGestureArea
The timer was used to let panning start with any movement after a
certain time. Before that time the startDragDistance from a platform
was used as threshold for pan start.
This causes trouble on platforms with touch interaction since start of
panning breaks a long press. When using touch before this change, a
long press on an item on top of the QDeclarativeGeoMapGestureArea
was difficult to achieve.

Change-Id: I1a265626b1b28dd5618eb83239a946ad8d25187a
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>