Remove the GeoMapMouseArea
I don't see a reason for keeping this. First of all, only half of it is
implemented. Second we can have the same behavior with a normal
MouseArea. The only advantage a MapMouseArea has is that mouse events
are only received in the boundaries of the parent map item. However
with a bit of work in the parent map item we can have this behavior as
well for a normal mouse area.

The only thing we are eventually loosing is the geocoordinate property
of the MapMouseEvent. But this can be easily obtained with the
"toCoordinate" function in the Map item.

Change-Id: Icb176ee7d7c2881df714ab3191fa1f7f5c8915e2
Reviewed-by: Aaron McCarthy <mccarthy.aaron@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>