Fix requesting single updates from Geoclue plugin.
Geoclue providers are expected to return a position update straight
away, otherwise the geoclue-master blocks when positioning is started.
This means that a position update with no longitude and latitude values
is not an error.

Change the Geoclue plugin to ignore these position updates, instead
waiting for the specified timeout before aborting.

Change-Id: I80b3c4cfdc31da78660f7435fdfac03963688c5a
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>