Fix declarative_ui tests.
Coordinates for map items located outside the view port cannot be
converted to and from screen coordinates. During tests pan the map so
that the items under tests are inside the view port.

Tests for panning the map and dragging map items were failing because
mapping between coordinates and screen position only works for visible
screen positions. The MouseAreas under test did not have
preventStealing set, which was causing the map to pan slightly prior to
item drag being detected.

Stop gesture detection for a touch point when the mouse is grabbed by a
child item.

Task-number: QTBUG-31797
Change-Id: I42a801d50c194c6fe9629f3b0de606b406ad8b8c
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>