Add direction and vertical speed properties to QML Position type.
The QML and C++ position types now provide the same functionality.

Position attributes are always set/reset on every position update. This
ensures that old values of speed, accuracy, direction, and vertical
speed are not indefinitely reported in QML. Values are set to NaN and if
appropriate the associated is valid property is set.

[ChangeLog][QtPositioning][Position] Added direction and vertical speed
properities.

Change-Id: Ia02da8a5bc1cd68489eb58180986df512c24aa2f
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>