Test that the mapping manager engine pointer is still valid.
QPointer holds a guarded pointer to a QObject. It must always be tested
prior to use.

Changed the parameter type to remove the need for the static_cast when
setting the QPointer member.

The root cause was that the tile fetcher object was not being destroyed
when the mapping manager engine was. Which was because the two objects
were in different threads. Some of the service specific tile fetchers
were directly calling into the engine objects across thread. Gah.

Since Qt 5 QNAM uses an application global thread for processing
requests, there is no need to use threads in the tile fetcher.

Change-Id: Id9df35ddfa78df2cbf334006fe5fc9726a75f92d
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>