Avoid excessive QImage::fill/copy in copyright handler
This fixes two problems. The first is that an empty copyright string
would result in the image being regenerated over and over again.
The second problem is that the code would allocate a fullscreen image
regardless of the size of the copyright watermark and then do
QImage::copy() to crop it to the displayed size. Needless to say,
this took a bit more time than strictly required.

Change-Id: Ia8d76e1cb34793439f92e7bb3ae00a8e99ffdc5e
Reviewed-by: Robin Burchell <robin+qt@viroteck.net>
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>