Ignore invalid center coordinates on maps
This causes undefined behavior and plenty of painter errors due to the
occurrence of NaN.

At the same time we change the default center to
London. Previously it was in the middle of the Atlantic. This led to
confusions because the all-blue display was mistaken for
nothing/empty/invalid tiles (which is represented by grey). By showing
the city of London we ensure that the default shows an easily
recognisable area.

Change-Id: I65a99d69d961b00a899120e190030e08a5ca0498
Reviewed-by: Robin Burchell <robin+qt@viroteck.net>
Reviewed-by: Gunnar Sletta <gunnar.sletta@jollamobile.com>