Build snippet outside of example build configurations too.
This is a partial revert of 9cbddccc24cdd4b51285ea9ecbf851f3eba2cd31.

The CI does not test a combined developer-build and example
configuration. Ensuring the snippet compiles is mostly a private test.
This patch drops the additional example condition to provide an
increased chance of compiling the snippets on the CI.

So far the condition never triggered and hence the CI couldn't ensure
a compiling snippet.

Task-number: QTBUG-37638
Change-Id: I8e64bb2b53b0452c5168faeb37f2809e96ac6e77
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>