Fix crash on S4 when running Satellite and Position updates concurrently
This was caused due to unsynchronized access of data between two threads.
The Java satellite updates were sent before the receiving Qt thread could
finish its regular startup. As a consequence the received updates could
not be mapped to a receiving C++ class.

This patch synchronizes the startup and setup of the two threads and
ensures that the relvant data setup on the C++ side is done before the
update thread is started.

In addition some minor debug helpers were added for future references.

Task-number: QTBUG-39082
Change-Id: I1f24ddfcef038fa000801ddb6a3ea0959370f0d3
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>