Fix QGeoRouteReplyOsm::networkReplyFinished
It can happen that the osm router does not find a route between the
specified points. In this case it returns a json object with a status
code other than 0 and an error message.
This commit notifies the app developer that an error occurred and
that no valid route is available.
Prior this it was not possible to catch the error message.

Change-Id: I172ae3cf9dfcd80e5a162110178203084071c9b0
Reviewed-by: Aaron McCarthy <mccarthy.aaron@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>