Allow HTML based copyright notices.
In addition to the image based map copyright/attribution notice allow a
geo service provider to provide a HTML string which is rendered and
displayed.

Add copyrightLinkActivated signal to the map to enable the application
to launch the url in an external browser.

The geo service providers are no longer in control of the position of
the copyright notice. For the time being it will be placed at the
bottom left of the map.

Change-Id: I49bfc58a70e9254220903d5103c714c08f327e8b
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>