Make QtLocation work after internal valuetype changes in QtQML
The value type wrapper code was completely removed. Instead
QGeoShape & derived classes and QGeoCoordinate use Q_GADGET
to advertise their meta data which in turn allows QtQML
to dynamically discover the types when needed.

This implies that the above C++ classes become larger as the meta
data is permanently added. On the positive side all QML value type
wrapper were removed which reduces the QML plugin size. At the
same time it becomes significantly easier to use those types in
other QML plugins such as the QtLocation one.

Unfortunately this change comes with an unavoidable behavior change
as well. Previously is was possible to set a width and height
on an invalid QML rectangle. In the case of an invalid rect it was
made valid (setting center to 0,0) before setting the height or
width. The QGeoRectangle class explicitly aborts calls to
setHeight() and setWidth() in case of an invalid rect. From
now on the C++ class behavior is used in QML too. This was
the reason for the change to tst_map_routing.qml in this patch.
Incidentally the documentation always referred to the QGeoRectangle
documentation and thus always stated the now new behavior.

Documentation for the value types and the relevant behavior changes
will be re-added in a later commit.

Each Gadget must be registered
using QMetaType::registerComparators<T>() to enable == and !=
operators in QML. Unfortunately QMetaType requires the operator<
which does not make any sense for the value types touched by
this patch. Right now, a warning is printed. Attempts will
be made to remove the operator< requirement in QMetaType.
Otherwise the warning has to be removed before the release
as they are printed very often.

[ChangeLog][QtPositioning][Important Behavior Changes] The QML
rectangle type ignores setting of the height and width if the
rectangle is invalid. Previously the property setter changed
the rectangle center to QGeoCoordinate(0,0) before setting the
height and width of a previously invalid rectangle.

Change-Id: Iab98ede58066ce03b2b4e0ce03ed85ba7ecb1020
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>