Don't override user's Map.zoomLevel when initializing QGeoCameraData
The user-chosen zoom level was reset during QGeoMap initialization.
This was caused by the construction of the QGeoMapController which in
turn constructed the first QGeoCameraData instance which had a default
initialised zoom level of 9. This default zoom level was immidiately
set as the maps zoom level which in turn invalidated the user's choice.

[ChangeLog][QtLocation][QML Map] Prevents overriding of user chosen
map zoom level during initialization.

Task-number: QTBUG-40779
Change-Id: I074632111738ffad041a7c54040168abca018fa2
Reviewed-by: Michal Klocek <michal.klocek@digia.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>