Use separate cache dir per geoservice/mapping plugin
This has the advantage that concurrent use of multiple plugins
doesn't force one plugin to load the tiles and queues of another plugin
into its tile cache.

QGeoTileCache could potentially separate the non-relevant tiles as well
but that's even more effort and requires the QGeoTileCache instance
to know to which QGeoMappingManagerEngine it belongs.

Task-number: QTBUG-40780
Change-Id: I342168c30a424b9f66d9c4e07694c1e3a9f2a412
Reviewed-by: Michal Klocek <michal.klocek@theqtcompany.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>