Add Q_DECL_OVERRIDE where missing
clang 3.6 and later print a warning for such cases. This patch fixes
all override warnings caused by the content of the qtlocation itself.
The remaining cases are triggered by headers outside of the qtlocation
repo.

This patch only fixes cases which clang on Linux can reach. It is
possible that some of the platform specific plugins still have some
ovveride related cases.

Change-Id: I9f93062511d38cba4449b7fc70e65c93d4251a8b
Reviewed-by: Michal Klocek <michal.klocek@theqtcompany.com>
Reviewed-by: Aaron McCarthy <mccarthy.aaron@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>