Fix duplicated link targets
The c++ to QML interaction document exists twice. Once for QtLocation
and a second time for QtPositioning. Although the content is different
the page title was same which created broken links.

The very same pages also created hard links with \target which were
named the same way the QML type was named. This created clashing
link targets.

Change-Id: Ie186217b11a606562a640b1e242fd73d8922327b
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>