Qualify positioning headers when including them from elsewhere
There are some corner cases in which you cannot rely on qmake
to generate module-specific include directives for the compiler
when building different parts of the same module. The rules on
this are complicated and it's easy to make mistakes.

Qualified includes are also locatable without module-specific
include directives. It's generally a good idea to qualify
includes from the same module but a different library in order
not to rely on qmake's magic too much.

Change-Id: I8703116c7c5c898394d7f393630cbfa655174605
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>