Fix build system
With the previous setup the location module wasn't properly set up
when qmake was run for the first time with '-r' and no previous
version of QtLocation installed. This caused the building of some
plugins to be skipped later.

It also just looks so much better with the dependencies and subdirs
properly grouped and without the ugly $$QT.location.include things.

Task-number: QTBUG-44983
Change-Id: I0396df078b755a1316ba896ece77620184286a86
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>