Improve Polyline MapItem
The old implementation used polygons added to a QPainterPath for the
screenBounds and the "contains" area calculation on each screen update
and thus had a huge performance impact when using more complex
polyline map items with several hundred points.
The new implementation only calculates the screenBounds on each screen
update. However, the contains check uses a similar logic to the
previous QPainterPath approach but is way more performant. This is
achieved by testing only parts of the polyline and not the whole
QPainterPath at once.

Change-Id: Ia332343f40ff865079fc86f0241a3f46a650162d
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>