Fix crash when initiating map flick w/o QtPositioning import
A map flick requires a QPropertyAnimation for a QGeoCoordinate. The
required animation interpolator is exported by the QtPositioning
QML plugin. In general a QML Location app should always import
QtPositioning as well. In this case the import was missing and
subsequently the interpolator for the QGeoCoordinate was never
registered.

This patch ensures that the QtLocation QML code that requires the
animation registers the interpolator in addition to the regular
QtPositioning registration.

Change-Id: I61759418f0cdf1cb30ff946cc325a112e90c1a2f
Task-number: QTBUG-46147
Reviewed-by: Michal Klocek <michal.klocek@theqtcompany.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>