Workaround for failing conversion between Q_GADGET's
QGeoRectangle and QGeoCircle are subclasses of QGeoShape. Before this
patch setting a QGeoShape property of a QML object is not possible
because the QGeoRectangle/QGeoCircle cannot be converted to a QGeoShape.
This patch addresses the problem by providing custom converters for
the above classes. It enables QVariant::convert() for the classes above.

Change-Id: I7daba5254dad0df598324432db3b1819e68648b6
Task-number: QTBUG-46232
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>
Reviewed-by: Michal Klocek <michal.klocek@theqtcompany.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>