Improve error handling in map, geo and route models
Current error handling is misleading and most of the error messages
are logged only to the console. Moreover error messages are too general.
For example 'plugin does not support mapping' is reported when
provider credentials are missing. Emitting onErrorChanged
and onErrorStringChanged does not work correctly, ie. check for
errorString in onErrorChanged ends up in not up to date error message.

Replace logging to console with setting an error on given component.
Use setError instead of qWarning and use one single setter
for errorType and errorString. Combine geocode error reporting
with geoprovider errors. Combine route error reporting with
geoprovider errors. Update unit tests. Move geocode unit test
to _core since it is not a _ui test.

Change-Id: Ia02b1bf4d88f48b963ada75fa7ed74598713c97f
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>