Fix disfunctional disk cache with OSM
The MapQuest servers use JPEG, not PNG. The disk cache, unlike the memory
one, passed in the extension as the format to loadFromData(). This failed
for JPEG files with an extension of .png.

In addition, using split('.')[1] to determine the file extension is bogus.
On Linux for example, a directory of "$HOME/.cache/QtLocation/..." results
in putting "cache/..." into format which is then conveniently ignored by
QImageReader. On OS X however there is no dot, so it "correctly" picks
"png" for the JPEG data, hence the console is flooded with warnings about
failed tile loading.

On top of the above, the disk cache is made more robust by not relying on the
format string. As long as only formats supported by QImageReader are in use,
it can recognize the correct format from the data. (the memory cache does this
already anyway)

Change-Id: I90ece5bd157d457afa0025527c84070427a60313
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>