Isolate the default cache directory
Instead of having it in several random places, have a central function
that contains all the cache directory selection logic.

As per QtCore docs GenericCacheLocation can be empty so introduce a fallback
to CacheLocation (the app-specific, non-shared location) in case some
future platform does not support the former.

Task-number: QTBUG-41187
Change-Id: Icfe5e8926e917483ac11844f625244318ff815c7
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>