Check if GenericCacheLocation is writable before trusting it
Enforced application isolation (AppArmor) may make it impossible to write
to $HOME/.cache as applications are only allowed to use their own specific
directories (as returned by CacheLocation). QtCore is not able to recognize
this so the value returned for GenericCacheLocation is still non-empty in
this case. Therefore do a write check.

Task-number: QTBUG-41187
Change-Id: I6ad136abc69997d25715dcdbbc1b8ffb55ec65e0
Reviewed-by: Alberto Mardegan <mardy@users.sourceforge.net>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>