Bump all QtLocation 5.5 tags to 5.6
5.5 was a tech preview release. The real release is 5.6.

Some 5.5 interfaces did change. It is possible that there are apps
in existence which are based on the 5.5 API's. To avoid miscommunication
we bump all version tags to the real first release version.

The existing 5.5 QML imports will continue to work but apps may encounter
API changes when they move from 5.5 to 5.6. The API changes are listed
in the QtLocation documentation.

Change-Id: Ia7000e871a5ee6b09234e5d58d4fc8927e93b154
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>