QtLocation mapviewer example MiniMap improvement
The mapviewer example currently uses the minimap to show a red
MapRectangle sized 10 degrees in latitude and 10 degrees in longitude.
This patch makes the minimap visualize the actual displayed area in
the main map as a red rectangle in the minimap.
The minimap zoom level is also now gently bound to the main map zoom
level.
There's one bug left, though, but not related to this qml component, which
causes the border of the red MapRectangle of the minimap to often be drawn
on the left of the rectangle.

Change-Id: I6ba332bf72f0489f5ada765c85f6890a09ca1a98
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>