Fix endless location requesting
This patch fixes the problem of lost positioning listeners, which was
registered in LocationManager, but does not persist in the `runningListeners`
collection. We can't call LocationManager.removeUpdates() for those listeners, that are not in `runningListeners`
and that's why LocationManager never ends to update location for them.

Listener can be registered in LocationManager and not added to
`runningListeners` collection at least in the following cases:

1. `startUpdates()` was called when location providers was disabled in
system.

2. `startUpdates()` was called twice, without `stopUpdates()`

Change-Id: If6777677cbb6f2ad9107fe2ac8496b7cdbb2d8d4
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>