Fix for crash on tessellation of self-intersectiong GeoPolygons
QtLocation makes use of a 3rd party library for polygon tessellation,
poly2tri.
This library is known to be weak in handling complex input, like, for
example, polygons with self intersecting edges.
Sanitizing the input data through clipper solves the problem, and it
has already been implemented in the clip2tri library, which this
patch includes and uses.

Task-number: QTBUG-52076
Change-Id: I071a58e202345bc71da583995f7245361f00e8c4
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>