Fix for wandering QGeoMapPolylineGeometry
QGeoMapPolylineGeometry is currently erratic due to the semi-randomly
varying geoLeftBound, that basically defines how the line will be drawn.
This patch uses a consistent way to calculate this value in all the
QDeclarativeGeo-items, at coordinate-definition time.

The new approach implies that each pair of points in a polyline
are connected in the longitudinal direction of the shortest path.
The patch attempts to optimize the common addCoordinate path by
incrementally searching for the left bound.

The patch also makes the other geoshapes (GeoPolygon, GeoCircle)
behave in the same way.

Task-number: QTBUG-43107
Task-number: QTBUG-52610
Change-Id: I7a0f6c4370fd6c50999fd2e5b1a7aa5954c8a8fc
Reviewed-by: Michal Klocek <michal.klocek@theqtcompany.com>