Polish and update map elements on window resize
When the a map element gets resized, the center stays the same, meaning
that map estate gets added (or removed) on each side of the element in
equal amount.
This means that the old screen coordinates of a map element change, and
therefore it must be repolished and updated.

Task-number: QTBUG-53128
Change-Id: Id26b111e8e33075082fa599e6eac92cb8089f5e4
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>