fix for losing map item focus while dragging them
QDeclarativeGeoMapItemBase::childMouseEventFilter currently does a
contain() check on the clicks but also on the mouse move events.
by not doing it on mouse move events the problem seems fixed, but the
autotests get broken because it seems that the map doesn't get panned
anymore but the element gets dragged instead.

Task-number: QTBUG-52075
Change-Id: I5ce340ca591ec924eebed1fe7acda18d31301fc5
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>