Delay removal/readd of map items until incubation completion on reset
When the model is reset, QDeclarativeGeoMapItemView currently removes
all map items and then begins recreating new delegates for model items.

This causes flickering since the mapitems are removed at once, but
the repopulation is asynchronous and allows the scene graph to repaint
in the meanwhile.

This patch delays the removal of current map items until all the new
items have been all incubated, so that removal and readd happen at
the same point.

The patch also adds additional checks to methods like
itemModelRowsRemoved, to verify if an item is still incubating.

Task-number: QTBUG-52301
Task-number: QTBUG-54188
Change-Id: I2f27114c3898bd61b1282b1f27b5f062bed1fe35
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>