Refactor: rename improperly named methods
After consultation, it seems that Qt favors using "At" rather than
"For" in naming a method such as "maximumLatitudeForZoom".

Moreover, this methods returns the maximum latitude that map center can
have at a specific zoom level, so it is important to make that clear
not only in the comments but also in the method's name.

Change-Id: I26f0d18f8ca94b1bdd085cd49dd603d51a68689b
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>